Organizes scene related classes together
