+++ StateMachine/Cpp/StateMachine/Scene.h
@@ -1,7 +1,26 @@
 #pragma once
+
+#ifndef SCENE_CONTROLLER_H
+#define SCENE_CONTROLLER_H
+
+class scene;
+
+class scene_controller
+{
+public:
+	scene_controller();
+	~scene_controller() = default;
+	void set_scene(scene* scene, bool deleteOldScene = false);
+	void update() const;
+private:
+	scene* scene_;
+};
+
+#endif
+
 #ifndef SCENE_H
 #define SCENE_H
-#include "scene_controller.h"
+#include <iostream>
 
 class scene_controller;
 
@@ -15,6 +34,38 @@ public:
 	virtual void end() = 0;
 protected:
 	scene_controller* controller_;
+};
+
+class start_scene : public scene
+{
+public:
+	start_scene(scene_controller* controller);
+	~start_scene() = default;
+
+	void start() override;
+	void update() override;
+	void end() override;
+};
+
+class menu_scene : public scene
+{
+public:
+	menu_scene(scene_controller* controller);
+	~menu_scene() = default;
+
+	void start() override;
+	void update() override;
+	void end() override;
+};
+
+class battle_scene : public scene
+{
+public:
+	battle_scene(scene_controller* controller);
+	~battle_scene() = default;
+	void start() override;
+	void update() override;
+	void end() override;
 };
 
 #endif
